Updated additional tests

+++ policyengine_us/variables/gov/states/nj/tax/income/exemptions/nj_blind_exemption.py
@@ -20,15 +20,15 @@ class nj_blind_exemption(Variable):
         p = parameters(period).gov.states.nj.tax.income.exemptions.blind
 
         # Get the individual blind status and disabled.
-        blind_head = tax_unit("blind_head", period).astype(int)
-        disabled_head = tax_unit("disabled_head", period).astype(int)
+        blind_head = tax_unit("blind_head", period)
+        disabled_head = tax_unit("disabled_head", period)
 
         # Check if the individual's eligiblity.
         head_eligible = blind_head | disabled_head
 
         # Get the individual's spouse blind status and disabled.
-        blind_spouse = (tax_unit("blind_spouse", period) * joint).astype(int)
-        disabled_spouse = tax_unit("disabled_spouse", period).astype(int)
+        blind_spouse = tax_unit("blind_spouse", period) * joint
+        disabled_spouse = tax_unit("disabled_spouse", period)
 
         # Check if the individual spouse's eligiblity.
         spouse_eligible = blind_spouse | disabled_spouse
